Remove the usage of Azure/Event-Hub (#739)
* Initial commit

* Restore upgrade_kyma_step_test.go, removed by mistake

* Bump KEB image to PR-739
