+++ resources/kcp/values.yaml
@@ -12,7 +12,7 @@ global:
       version: "PR-730"
     kyma_environment_broker:
       dir:
-      version: "PR-738"
+      version: "PR-739"
     kyma_environments_cleanup_job:
       dir:
       version: "PR-723"
